+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change Log
 
+## Version
+
+1. Added: Ability to return new query of anynomous alias function in from() method
+
 ## Version 3.2.0
 
 1. Added: Ability to query from table alias